Limit expansion when size is reduced
The idea of this change is to try to constrain how much padding
a sender can add.  This is done by calculating the size of a
record with maximum plaintext and minimum padding.  No other record
can exceed this size, though smaller packets can pad to reach this
size.

I was going to include a formula, then realized that the simple
answer is best: pad to the record size limit, then pad to get
to a multiple of the block size as normal.  Most implementations
will be OK because they don't bother with the first step.