Move the AEAD justification text up
This is the primary motivation for the draft and it should be more prominent.

I've made this more general, noting that processing and releasing incomplete
records exposes a forgery risk.  There are probably many other concerns as
well, such as side-channel risks (what if you have a partial block, are you
*sure* that you implemented the partial release without leaving a timing
side-channel?), but I think that the forgery concern is sufficiently
motivating.