Relationship with other constraints on implementations
Add an applicability statement.  This only addresses buffer space issues, not
processing or bandwidth constraints.  I didn't observe (in the introduction)
that this can negatively affect bandwidth, that's later.

Hannes also observed that we don't really talk about the path MTU.  I made it
clear that these are separate and independent limitations.