Allow DTLS endpoints to discard big records
The current text doesn't work for DTLS, where receipt of a UDP datagram
exceeding the record size should not be automatically fatal to the connection.
Such packets can be discarded: a record should only cause a fatal alert if it
is decrypted successfully.  And most implementations won't want to commit to
decryption of records that are too big.